feat(cart): add logic cart without login

+++ ecommerce/app/Http/Livewire/Cart.php
@@ -34,31 +34,67 @@ class Cart extends Component
 
     public function render()
     {
-        $this->carts = DB::table('carts')
-            ->join('products', 'carts.product_id', 'products.id')
-            ->leftJoin('detail_products', function ($join) {
-                $join->on('carts.product_id', 'detail_products.dp_id')
-                    ->whereColumn('detail_products.size', 'carts.size');
-            })
-            ->where('carts.user_id', Auth::id())
-            ->select(
-                'carts.id AS CartID',
-                'products.id AS ProductID',
-                'products.product_id',
-                'products.name AS ProdName',
-                DB::raw('COALESCE(detail_products.stock, 0) AS AvailStock'),
-                'products.image',
-                DB::raw('products.price * carts.qty AS price'),
-                'carts.size',
-                'carts.qty'
-            )
-            ->get();
+        if (Auth::check()) {
+            $this->carts = DB::table('carts')
+                ->join('products', 'carts.product_id', 'products.id')
+                ->leftJoin('detail_products', function ($join) {
+                    $join->on('carts.product_id', 'detail_products.dp_id')
+                        ->whereColumn('detail_products.size', 'carts.size');
+                })
+                ->where('carts.user_id', Auth::id())
+                ->select(
+                    'carts.id AS CartID',
+                    'products.id AS ProductID',
+                    'products.product_id',
+                    'products.name AS ProdName',
+                    DB::raw('COALESCE(detail_products.stock, 0) AS AvailStock'),
+                    'products.image',
+                    DB::raw('products.price * carts.qty AS price'),
+                    'carts.size',
+                    'carts.qty'
+                )
+                ->get();
 
-        $this->availStock = $this->carts->filter(fn ($cart) => $cart->AvailStock);
+            $this->availStock = $this->carts->filter(fn ($cart) => $cart->AvailStock);
+        }
+
+        if (isset($_COOKIE['cart_id']) && isset($_COOKIE['carts'])) {
+            $dataArray = json_decode($_COOKIE['carts'], true);
+
+            $results = DB::table('products')
+                ->join('detail_products', 'products.id', '=', 'detail_products.dp_id')
+                ->select(
+                    'products.id AS ProductID',
+                    'products.product_id',
+                    'products.name AS ProdName',
+                    'products.image',
+                    DB::raw('COALESCE(detail_products.stock, 0) AS AvailStock'),
+                    DB::raw('products.price AS price'),
+                    DB::raw('products.price * detail_products.stock AS total_price'),
+                    'detail_products.size'
+                )
+                ->whereIn('detail_products.dp_id', array_column($dataArray, 'product_id'))
+                ->whereIn('detail_products.size', array_column($dataArray, 'size'))
+                ->get();
+
+            // Processing and formatting the results...
+            foreach ($results as &$result) {
+                $productId = $result->ProductID;
+
+                foreach ($dataArray as $item) {
+                    if ($item['product_id'] == $productId) {
+                        $result->CartID = $item['cart_id'];
+                        $result->qty = $item['quantity'];
+                        break;
+                    }
+                }
+            }
+
+            $this->carts = $results;
+        }
 
         return view('livewire.cart');
     }
-
 
     public function initProp($cartID, $productID)
     {
@@ -88,23 +124,40 @@ class Cart extends Component
 
     public function removeCartItem()
     {
-        $cart = CartModel::where('id', $this->cartID)
-            ->where('user_id', Auth::id())
-            ->where('product_id', $this->productID)
-            ->first();
+        if (Auth::check()) {
+            $cart = CartModel::where('id', $this->cartID)
+                ->where('user_id', Auth::id())
+                ->where('product_id', $this->productID)
+                ->first();
 
-        if ($cart) {
-            $cart->delete();
+            if ($cart) {
+                $cart->delete();
 
-            $this->emit('refreshTotalPrice');
+                $this->emit('refreshTotalPrice');
+                $this->emit('refreshCart');
+
+                return redirect($this->page)->with('status', 200);
+            } else {
+                return $this->dispatchBrowserEvent('toastr', [
+                    'status' => 'error',
+                    'message' => 'Gagal menghapus pesanan'
+                ]);
+            }
+        }
+
+        if (!Auth::check() && isset($_COOKIE['cart_id']) && isset($_COOKIE['carts'])) {
+            $dataArray = json_decode($_COOKIE['carts'], true);
+
+            $newArray = [];
+            foreach ($dataArray as $item) {
+                if ($item['cart_id'] != $this->cartID) {
+                    $newArray[] = $item;
+                }
+            }
+
             $this->emit('refreshCart');
-
+            setcookie('carts', json_encode($newArray), time() + (3600 * 2), '/');
             return redirect($this->page)->with('status', 200);
-        } else {
-            return $this->dispatchBrowserEvent('toastr', [
-                'status' => 'error',
-                'message' => 'Gagal menghapus pesanan'
-            ]);
         }
     }
